backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-genai
pytest
pytest-asyncio
pytest-mock
httpx
